Add slideshow of the car list
